+++ source/common/ipc/api.js
@@ -38,6 +38,10 @@ import type {
   ResumeDownloadResponse,
 } from '../types/downloadManager.types';
 import type { StoreMessage } from '../types/electron-store.types';
+import type {
+  IntrospectAddressRequest,
+  IntrospectAddressResponse,
+} from '../types/address-introspection.types';
 
 /**
  * ======================= IPC CHANNELS API =========================
@@ -320,6 +324,13 @@ export type DownloadMainResponse = DownloadResponse;
 export const RESUME_DOWNLOAD = 'RESUME_DOWNLOAD';
 export type ResumeDownloadRendererRequest = ResumeDownloadRequest;
 export type ResumeDownloadMainResponse = ResumeDownloadResponse | void;
+
+/**
+ * Channel for introspecting an address
+ */
+export const INTROSPECT_ADDRESS_CHANNEL = 'INTROSPECT_ADDRESS_CHANNEL';
+export type IntrospectAddressRendererRequest = IntrospectAddressRequest;
+export type IntrospectAddressMainResponse = IntrospectAddressResponse;
 
 /**
  * Channel for checking block replay progress