+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Changelog
 
 ### Features
 
-- Show block replay progress ([PR 2087](https://github.com/input-output-hk/daedalus/pull/2087))
+- Show block replay progress ([PR 2110](https://github.com/input-output-hk/daedalus/pull/2110))
 
 ### Fixes